--- a/tests/test_language_sentences.py
+++ b/tests/test_language_sentences.py
@@ -28,6 +28,8 @@ def do_test_language_sentences_file(
     """Tests recognition all of the test sentences for a language"""
     _testing_domain, testing_intent = test_file.split("_", 1)
 
+    seen_sentences = set()
+
     for test in load_test(language, test_file)["tests"]:
         intent = test["intent"]
         assert (
@@ -35,6 +37,11 @@ def do_test_language_sentences_file(
         ), f"File {test_file} should only test for intent {testing_intent}"
 
         for sentence in test["sentences"]:
+            assert (
+                sentence not in seen_sentences
+            ), f"Duplicate sentence found: {sentence}"
+            seen_sentences.add(sentence)
+
             result = recognize(sentence, language_sentences, slot_lists=slot_lists)
             assert result is not None, f"Recognition failed for '{sentence}'"
             assert (