@balloob 2022-12-25
Caching the parsing of `_common.yaml` solved everything.

```
288 passed in 11.60s

Slowest:
1.49s setup    tests/test_language_intents.py::test_homeassistant_HassTurnOn[de]
1.00s call     tests/test_language_intents.py::test_light_HassLightSet[hu]
0.75s call     tests/test_language_intents.py::test_cover_HassOpenCover[de]
```

There is now no noticeable speed loss per test.